+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Expose known block height of peer in brand new `height` field added in responses of `/network/connections` API endpoints
 - `-verify-db` option (default true), will verify the database integrity during startup and exit if a problem is found
 - `-reset-corrupt-db` option (default false) will verify the database integrity during startup and reset the db if a problem is found
+- `GET /explorer/address`: add `size` and `fee` to transaction objects and `calculated_hours` to transaction inputs
 
 ### Fixed